Split NDJSON rows planning
