+++ jetro-core/src/io/mod.rs
@@ -13,6 +13,7 @@ mod ndjson_frame;
 mod ndjson_hint;
 mod ndjson_parallel;
 mod ndjson_rev;
+mod ndjson_rows;
 mod ndjson_scan;
 mod ndjson_stream_cache;
 mod source;
